requirements.txt
pygame 
py-trees 
pyyaml
imageio